+++ transports/dns/src/lib.rs
@@ -24,10 +24,11 @@
 //! [`DnsConfig`] and `TokioDnsConfig` for use with `async-std` and `tokio`,
 //! respectively.
 //!
-//! A [`GenDnsConfig`] is a [`Transport`] wrapper that is created around
+//! A [`GenDnsConfig`] is an address-rewriting [`Transport`] wrapper around
 //! an inner `Transport`. The composed transport behaves like the inner
-//! transport, except that [`Transport::dial`] resolves `/dns`, `/dns4/` and
-//! `/dns6/` components of a given `Multiaddr` through a DNS.
+//! transport, except that [`Transport::dial`] resolves `/dns/...`, `/dns4/...`,
+//! `/dns6/...` and `/dnsaddr/...` components of the given `Multiaddr` through
+//! a DNS, replacing them with the resolved protocols (typically TCP/IP).
 //!
 //! The `async-std` feature and hence the `DnsConfig` are
 //! enabled by default. Tokio users can furthermore opt-in
@@ -37,14 +38,14 @@
 //!
 //![trust-dns-resolver]: https://docs.rs/trust-dns-resolver/latest/trust_dns_resolver/#dns-over-tls-and-dns-over-https
 
-use futures::{prelude::*, future::BoxFuture, stream::FuturesOrdered};
+use futures::{prelude::*, future::BoxFuture};
 use libp2p_core::{
     Transport,
     multiaddr::{Protocol, Multiaddr},
     transport::{TransportError, ListenerEvent}
 };
-use log::{debug, trace};
-use std::{error, fmt, net::IpAddr};
+use smallvec::SmallVec;
+use std::{borrow::Cow, convert::TryFrom, error, fmt, iter, net::IpAddr, str};
 #[cfg(any(feature = "async-std", feature = "tokio"))]
 use std::io;
 #[cfg(any(feature = "async-std", feature = "tokio"))]
@@ -61,6 +62,24 @@ use async_std_resolver::{AsyncStdConnection, AsyncStdConnectionProvider};
 
 pub use trust_dns_resolver::config::{ResolverConfig, ResolverOpts};
 pub use trust_dns_resolver::error::{ResolveError, ResolveErrorKind};
+
+/// The prefix for `dnsaddr` protocol TXT record lookups.
+const DNSADDR_PREFIX: &'static str = "_dnsaddr.";
+
+/// The maximum number of dialing attempts to resolved addresses.
+const MAX_DIAL_ATTEMPTS: usize = 16;
+
+/// The maximum number of DNS lookups when dialing.
+///
+/// This limit is primarily a safeguard against too many, possibly
+/// even cyclic, indirections in the addresses obtained from the
+/// TXT records of a `/dnsaddr`.
+const MAX_DNS_LOOKUPS: usize = 32;
+
+/// The maximum number of TXT records applicable for the address
+/// being dialed that are considered for further lookups as a
+/// result of a single `/dnsaddr` lookup.
+const MAX_TXT_RECORDS: usize = 16;
 
 /// A `Transport` wrapper for performing DNS lookups when dialing `Multiaddr`esses
 /// using `async-std` for all async I/O.
@@ -137,7 +156,7 @@ where
 
 impl<T, C, P> Transport for GenDnsConfig<T, C, P>
 where
-    T: Transport + Send + 'static,
+    T: Transport + Clone + Send + 'static,
     T::Error: Send,
     T::Dial: Send,
     C: DnsHandle<Error = ResolveError>,
@@ -171,44 +190,120 @@ where
     }
 
     fn dial(self, addr: Multiaddr) -> Result<Self::Dial, TransportError<Self::Error>> {
-        // Check if there are any domain names in the address. If not, proceed
-        // straight away with dialing on the underlying transport.
-        if !addr.iter().any(|p| match p {
-            Protocol::Dns(_) | Protocol::Dns4(_) | Protocol::Dns6(_) => true,
-            _ => false
-        }) {
-            trace!("Pass-through address without DNS: {}", addr);
-            let inner_dial = self.inner.dial(addr)
-                .map_err(|err| err.map(DnsErr::Transport))?;
-            return Ok(inner_dial.map_err::<_, fn(_) -> _>(DnsErr::Transport).left_future());
-        }
-
         // Asynchronlously resolve all DNS names in the address before proceeding
         // with dialing on the underlying transport.
         Ok(async move {
             let resolver = self.resolver;
             let inner = self.inner;
 
-            trace!("Resolving DNS: {}", addr);
+            let mut last_err = None;
+            let mut dns_lookups = 0;
+            let mut dial_attempts = 0;
+            // We optimise for the common case of a single DNS component
+            // in the address that is resolved with a single lookup.
+            let mut unresolved = SmallVec::<[Multiaddr; 1]>::new();
+            unresolved.push(addr.clone());
 
-            let resolved = addr.into_iter()
-                .map(|proto| resolve(proto, &resolver))
-                .collect::<FuturesOrdered<_>>()
-                .collect::<Vec<Result<Protocol<'_>, Self::Error>>>()
-                .await
-                .into_iter()
-                .collect::<Result<Vec<Protocol<'_>>, Self::Error>>()?
-                .into_iter()
-                .collect::<Multiaddr>();
+            // Resolve (i.e. replace) all DNS protocol components, initiating
+            // dialing attempts as soon as there is another fully resolved
+            // address.
+            while let Some(addr) = unresolved.pop() {
+                if let Some((i, name)) = addr.iter().enumerate().find(|(_, p)| match p {
+                    Protocol::Dns(_) |
+                    Protocol::Dns4(_) |
+                    Protocol::Dns6(_) |
+                    Protocol::Dnsaddr(_) => true,
+                    _ => false
+                }) {
+                    if dns_lookups == MAX_DNS_LOOKUPS {
+                        log::debug!("Too many DNS lookups. Dropping unresolved {}.", addr);
+                        last_err = Some(DnsErr::TooManyLookups);
+                        // There may still be fully resolved addresses in `unresolved`,
+                        // so keep going until `unresolved` is empty.
+                        continue
+                    }
+                    dns_lookups += 1;
+                    match resolve(&name, &resolver).await {
+                        Err(e) => {
+                            if unresolved.is_empty() {
+                                return Err(e)
+                            }
+                            // If there are still unresolved addresses, there is
+                            // a chance of success, but we track the last error.
+                            last_err = Some(e);
+                        }
+                        Ok(Resolved::One(ip)) => {
+                            log::trace!("Resolved {} -> {}", name, ip);
+                            let addr = addr.replace(i, |_| Some(ip)).expect("`i` is a valid index");
+                            unresolved.push(addr);
+                        }
+                        Ok(Resolved::Many(ips)) => {
+                            for ip in ips {
+                                log::trace!("Resolved {} -> {}", name, ip);
+                                let addr = addr.replace(i, |_| Some(ip)).expect("`i` is a valid index");
+                                unresolved.push(addr);
+                            }
+                        }
+                        Ok(Resolved::Addrs(addrs)) => {
+                            let suffix = addr.iter().skip(i + 1).collect::<Multiaddr>();
+                            let prefix = addr.iter().take(i).collect::<Multiaddr>();
+                            let mut n = 0;
+                            for a in addrs {
+                                if a.ends_with(&suffix) {
+                                    if n < MAX_TXT_RECORDS {
+                                        n += 1;
+                                        log::trace!("Resolved {} -> {}", name, a);
+                                        let addr = prefix.iter().chain(a.iter()).collect::<Multiaddr>();
+                                        unresolved.push(addr);
+                                    } else {
+                                        log::debug!("Too many TXT records. Dropping resolved {}.", a);
+                                    }
+                                }
+                            }
+                        }
+                    }
+                } else {
+                    // We have a fully resolved address, so try to dial it.
+                    log::debug!("Dialing {}", addr);
 
-            debug!("DNS resolved: {} => {}", addr, resolved);
+                    let transport = inner.clone();
+                    let result = match transport.dial(addr) {
+                        Ok(out) => {
+                            // We only count attempts that the inner transport
+                            // actually accepted, i.e. for which it produced
+                            // a dialing future.
+                            dial_attempts += 1;
+                            out.await.map_err(DnsErr::Transport)
+                        }
+                        Err(TransportError::MultiaddrNotSupported(a)) =>
+                            Err(DnsErr::MultiaddrNotSupported(a)),
+                        Err(TransportError::Other(err)) => Err(DnsErr::Transport(err))
+                    };
 
-            match inner.dial(resolved) {
-                Ok(out) => out.await.map_err(DnsErr::Transport),
-                Err(TransportError::MultiaddrNotSupported(a)) =>
-                    Err(DnsErr::MultiaddrNotSupported(a)),
-                Err(TransportError::Other(err)) => Err(DnsErr::Transport(err))
+                    match result {
+                        Ok(out) => return Ok(out),
+                        Err(err) => {
+                            log::debug!("Dial error: {:?}.", err);
+                            if unresolved.is_empty() {
+                                return Err(err)
+                            }
+                            if dial_attempts == MAX_DIAL_ATTEMPTS {
+                                log::debug!("Aborting dialing after {} attempts.", MAX_DIAL_ATTEMPTS);
+                                return Err(err)
+                            }
+                            last_err = Some(err);
+                        }
+                    }
+                }
             }
+
+            // At this point, if there was at least one failed dialing
+            // attempt, return that error. Otherwise there were no valid DNS records
+            // for the given address to begin with (i.e. DNS lookups succeeded but
+            // produced no records relevant for the given `addr`).
+            Err(last_err.unwrap_or_else(||
+                DnsErr::ResolveError(
+                    ResolveErrorKind::Message("No matching records found.").into())))
         }.boxed().right_future())
     }
 
@@ -226,6 +321,13 @@ pub enum DnsErr<TErr> {
     ResolveError(ResolveError),
     /// DNS resolution was successful, but the underlying transport refused the resolved address.
     MultiaddrNotSupported(Multiaddr),
+    /// DNS resolution involved too many lookups.
+    ///
+    /// DNS resolution on dialing performs up to 32 DNS lookups. If these
+    /// are not sufficient to obtain a fully-resolved address, this error
+    /// is returned and the DNS records for the domain(s) being dialed
+    /// should be investigated.
+    TooManyLookups,
 }
 
 impl<TErr> fmt::Display for DnsErr<TErr>
@@ -236,6 +338,7 @@ where TErr: fmt::Display
             DnsErr::Transport(err) => write!(f, "{}", err),
             DnsErr::ResolveError(err) => write!(f, "{}", err),
             DnsErr::MultiaddrNotSupported(a) => write!(f, "Unsupported resolved address: {}", a),
+            DnsErr::TooManyLookups => write!(f, "Too many DNS lookups"),
         }
     }
 }
@@ -248,14 +351,33 @@ where TErr: error::Error + 'static
             DnsErr::Transport(err) => Some(err),
             DnsErr::ResolveError(err) => Some(err),
             DnsErr::MultiaddrNotSupported(_) => None,
+            DnsErr::TooManyLookups => None,
         }
     }
 }
 
-/// Asynchronously resolves the domain name of a `Dns`, `Dns4` or `Dns6` protocol
-/// component. If the given protocol is not a DNS component, it is returned unchanged.
-fn resolve<'a, E: 'a, C, P>(proto: Protocol<'a>, resolver: &'a AsyncResolver<C,P>)
-    -> impl Future<Output = Result<Protocol<'a>, DnsErr<E>>> + 'a
+/// The successful outcome of [`resolve`] for a given [`Protocol`].
+enum Resolved<'a> {
+    /// The given `Protocol` has been resolved to a single `Protocol`,
+    /// which may be identical to the one given, in case it is not
+    /// a DNS protocol component.
+    One(Protocol<'a>),
+    /// The given `Protocol` has been resolved to multiple alternative
+    /// `Protocol`s as a result of a DNS lookup.
+    Many(Vec<Protocol<'a>>),
+    /// The given `Protocol` has been resolved to a new list of `Multiaddr`s
+    /// obtained from DNS TXT records representing possible alternatives.
+    /// These addresses may contain further DNS names that need resolving.
+    Addrs(Vec<Multiaddr>),
+}
+
+/// Asynchronously resolves the domain name of a `Dns`, `Dns4`, `Dns6` or `Dnsaddr` protocol
+/// component. If the given protocol is of a different type, it is returned unchanged as a
+/// [`Resolved::One`].
+fn resolve<'a, E: 'a + Send, C, P>(
+    proto: &Protocol<'a>,
+    resolver: &'a AsyncResolver<C,P>,
+) -> BoxFuture<'a, Result<Resolved<'a>, DnsErr<E>>>
 where
     C: DnsHandle<Error = ResolveError>,
     P: ConnectionProvider<Conn = C>,
@@ -263,39 +385,105 @@ where
     match proto {
         Protocol::Dns(ref name) => {
             resolver.lookup_ip(fqdn(name)).map(move |res| match res {
-                Ok(ips) => Ok(ips.into_iter()
-                    .next()
-                    .map(Protocol::from)
-                    .expect("If there are no results, `Err(NoRecordsFound)` is expected.")),
-                Err(e) => return Err(DnsErr::ResolveError(e))
-            }).left_future()
+                Ok(ips) => {
+                    let mut ips = ips.into_iter();
+                    let one = ips.next()
+                        .expect("If there are no results, `Err(NoRecordsFound)` is expected.");
+                    if let Some(two) = ips.next() {
+                        Ok(Resolved::Many(
+                            iter::once(one).chain(iter::once(two))
+                                .chain(ips)
+                                .map(Protocol::from)
+                                .collect()))
+                    } else {
+                        Ok(Resolved::One(Protocol::from(one)))
+                    }
+                }
+                Err(e) => Err(DnsErr::ResolveError(e))
+            }).boxed()
         }
         Protocol::Dns4(ref name) => {
             resolver.ipv4_lookup(fqdn(name)).map(move |res| match res {
-                Ok(ips) => Ok(ips.into_iter()
-                    .map(IpAddr::from)
-                    .next()
-                    .map(Protocol::from)
-                    .expect("If there are no results, `Err(NoRecordsFound)` is expected.")),
-                Err(e) => return Err(DnsErr::ResolveError(e))
-            }).left_future().left_future().right_future()
+                Ok(ips) => {
+                    let mut ips = ips.into_iter();
+                    let one = ips.next()
+                        .expect("If there are no results, `Err(NoRecordsFound)` is expected.");
+                    if let Some(two) = ips.next() {
+                        Ok(Resolved::Many(
+                            iter::once(one).chain(iter::once(two))
+                                .chain(ips)
+                                .map(IpAddr::from)
+                                .map(Protocol::from)
+                                .collect()))
+                    } else {
+                        Ok(Resolved::One(Protocol::from(IpAddr::from(one))))
+                    }
+                }
+                Err(e) => Err(DnsErr::ResolveError(e))
+            }).boxed()
         }
         Protocol::Dns6(ref name) => {
             resolver.ipv6_lookup(fqdn(name)).map(move |res| match res {
-                Ok(ips) => Ok(ips.into_iter()
-                    .map(IpAddr::from)
-                    .next()
-                    .map(Protocol::from)
-                    .expect("If there are no results, `Err(NoRecordsFound)` is expected.")),
-                Err(e) => return Err(DnsErr::ResolveError(e))
-            }).right_future().left_future().right_future()
+                Ok(ips) => {
+                    let mut ips = ips.into_iter();
+                    let one = ips.next()
+                        .expect("If there are no results, `Err(NoRecordsFound)` is expected.");
+                    if let Some(two) = ips.next() {
+                        Ok(Resolved::Many(
+                            iter::once(one).chain(iter::once(two))
+                                .chain(ips)
+                                .map(IpAddr::from)
+                                .map(Protocol::from)
+                                .collect()))
+                    } else {
+                        Ok(Resolved::One(Protocol::from(IpAddr::from(one))))
+                    }
+                }
+                Err(e) => Err(DnsErr::ResolveError(e))
+            }).boxed()
         },
-        proto => future::ready(Ok(proto)).right_future().right_future()
+        Protocol::Dnsaddr(ref name) => {
+            let name = Cow::Owned([DNSADDR_PREFIX, name].concat());
+            resolver.txt_lookup(fqdn(&name)).map(move |res| match res {
+                Ok(txts) => {
+                    let mut addrs = Vec::new();
+                    for txt in txts {
+                        if let Some(chars) = txt.txt_data().first() {
+                            match parse_dnsaddr_txt(chars) {
+                                Err(e) => {
+                                    // Skip over seemingly invalid entries.
+                                    log::debug!("Invalid TXT record: {:?}", e);
+                                }
+                                Ok(a) => {
+                                    addrs.push(a);
+                                }
+                            }
+                        }
+                    }
+                    Ok(Resolved::Addrs(addrs))
+                }
+                Err(e) => Err(DnsErr::ResolveError(e))
+            }).boxed()
+        }
+        proto => future::ready(Ok(Resolved::One(proto.clone()))).boxed()
     }
 }
 
-fn fqdn(name: &std::borrow::Cow<'_, str>) -> String {
-    if name.ends_with(".") {
+/// Parses a `<character-string>` of a `dnsaddr` TXT record.
+fn parse_dnsaddr_txt(txt: &[u8]) -> io::Result<Multiaddr> {
+    let s = str::from_utf8(txt).map_err(invalid_data)?;
+    match s.strip_prefix("dnsaddr=") {
+        None => Err(invalid_data("Missing `dnsaddr=` prefix.")),
+        Some(a) => Ok(Multiaddr::try_from(a).map_err(invalid_data)?)
+    }
+}
+
+fn invalid_data(e: impl Into<Box<dyn std::error::Error + Send + Sync>>) -> io::Error {
+    io::Error::new(io::ErrorKind::InvalidData, e)
+}
+
+fn fqdn(name: &Cow<'_, str>) -> String {
+    if name.ends_with('.') {
         name.to_string()
     } else {
         format!("{}.", name)
@@ -308,6 +496,7 @@ mod tests {
     use futures::{future::BoxFuture, stream::BoxStream};
     use libp2p_core::{
         Transport,
+        PeerId,
         multiaddr::{Protocol, Multiaddr},
         transport::ListenerEvent,
         transport::TransportError,
@@ -332,17 +521,12 @@ mod tests {
             }
 
             fn dial(self, addr: Multiaddr) -> Result<Self::Dial, TransportError<Self::Error>> {
-                let addr = addr.iter().collect::<Vec<_>>();
-                assert_eq!(addr.len(), 2);
-                match addr[1] {
-                    Protocol::Tcp(_) => (),
-                    _ => panic!(),
-                };
-                match addr[0] {
-                    Protocol::Ip4(_) => (),
-                    Protocol::Ip6(_) => (),
-                    _ => panic!(),
-                };
+                // Check that all DNS components have been resolved, i.e. replaced.
+                assert!(!addr.iter().any(|p| match p {
+                    Protocol::Dns(_) | Protocol::Dns4(_) | Protocol::Dns6(_) | Protocol::Dnsaddr(_)
+                        => true,
+                    _ => false,
+                }));
                 Ok(Box::pin(future::ready(Ok(()))))
             }
 
@@ -383,6 +567,37 @@ mod tests {
                 .await
                 .unwrap();
 
+            // Success due to the DNS TXT records at _dnsaddr.bootstrap.libp2p.io.
+            let _ = transport
+                .clone()
+                .dial("/dnsaddr/bootstrap.libp2p.io".parse().unwrap())
+                .unwrap()
+                .await
+                .unwrap();
+
+            // Success due to the DNS TXT records at _dnsaddr.bootstrap.libp2p.io having
+            // an entry with suffix `/p2p/QmNnooDu7bfjPFoTZYxMNLWUQJyrVwtbZg5gBMjTezGAJN`,
+            // i.e. a bootnode with such a peer ID.
+            let _ = transport
+                .clone()
+                .dial("/dnsaddr/bootstrap.libp2p.io/p2p/QmNnooDu7bfjPFoTZYxMNLWUQJyrVwtbZg5gBMjTezGAJN".parse().unwrap())
+                .unwrap()
+                .await
+                .unwrap();
+
+            // Failure due to the DNS TXT records at _dnsaddr.libp2p.io not having
+            // an entry with a random `p2p` suffix.
+            match transport
+                .clone()
+                .dial(format!("/dnsaddr/bootstrap.libp2p.io/p2p/{}", PeerId::random()).parse().unwrap())
+                .unwrap()
+                .await
+            {
+                Err(DnsErr::ResolveError(_)) => {},
+                Err(e) => panic!("Unexpected error: {:?}", e),
+                Ok(_) => panic!("Unexpected success.")
+            }
+
             // Failure due to no records.
             match transport
                 .clone()
@@ -401,19 +616,27 @@ mod tests {
 
         #[cfg(feature = "async-std")]
         {
+            // Be explicit about the resolver used. At least on github CI, TXT
+            // type record lookups may not work with the system DNS resolver.
+            let config = ResolverConfig::quad9();
+            let opts = ResolverOpts::default();
             async_std_crate::task::block_on(
-                DnsConfig::system(CustomTransport).then(|dns| run(dns.unwrap()))
+                DnsConfig::custom(CustomTransport, config, opts).then(|dns| run(dns.unwrap()))
             );
         }
 
         #[cfg(feature = "tokio")]
         {
+            // Be explicit about the resolver used. At least on github CI, TXT
+            // type record lookups may not work with the system DNS resolver.
+            let config = ResolverConfig::quad9();
+            let opts = ResolverOpts::default();
             let rt = tokio_crate::runtime::Builder::new_current_thread()
                 .enable_io()
                 .enable_time()
                 .build()
                 .unwrap();
-            rt.block_on(run(TokioDnsConfig::system(CustomTransport).unwrap()));
+            rt.block_on(run(TokioDnsConfig::custom(CustomTransport, config, opts).unwrap()));
         }
     }
 }