+++ transports/wasm-ext/CHANGELOG.md
@@ -1,3 +1,7 @@
+# 0.28.0 [unreleased]
+
+- Update `libp2p-core`.
+
 # 0.27.0 [2021-01-12]
 
 - Update dependencies.